requirements.txt
keras==2.10.0
Keras-Preprocessing==1.1.2
matplotlib==3.5.2
numpy==1.21.5
opencv-python-headless==4.6.0.66
pandas==1.4.3
scikit-learn==1.1.1
tensorflow==2.10.0
pillow~=9.2.0